Combine enrollment data and madgrades under term
