+++ generation/course.py
@@ -1,8 +1,6 @@
-import asyncio
 import re
-from json import JSONDecodeError
 
-from enrollment_data import EnrollmentData
+from enrollment_data import EnrollmentData, GradeData, TermData
 from json_serializable import JsonSerializable
 
 
@@ -109,16 +107,16 @@ class Course(JsonSerializable):
             description: str,
             prerequisites: Prerequisites,
             optimized_prerequisites: Prerequisites | None,
-            madgrades_data,
-            enrollment_data: dict[str, EnrollmentData],
+            cumulative_grade_data: "GradeData | None",
+            term_data: dict[str, TermData],
     ):
         self.course_reference = course_reference
         self.course_title = course_title
         self.description = description
         self.prerequisites = prerequisites
         self.optimized_prerequisites = optimized_prerequisites
-        self.madgrades_data = madgrades_data
-        self.enrollment_data = enrollment_data
+        self.cumulative_grade_data = cumulative_grade_data
+        self.term_data = term_data
 
     @classmethod
     def from_json(cls, json_data) -> "Course":
@@ -126,9 +124,9 @@ class Course(JsonSerializable):
         if json_data["optimized_prerequisites"]:
             optimized_prerequisites = Course.Prerequisites.from_json(json_data["optimized_prerequisites"])
 
-        madgrades_data = None
-        if json_data["madgrades_data"]:
-            madgrades_data = MadgradesData.from_json(json_data["madgrades_data"])
+        cumulative_grade_data = None
+        if json_data["cumulative_grade_data"]:
+            cumulative_grade_data = GradeData.from_json(json_data["cumulative_grade_data"])
 
         return Course(
             course_reference=Course.Reference.from_json(json_data["course_reference"]),
@@ -136,8 +134,8 @@ class Course(JsonSerializable):
             description=json_data["description"],
             prerequisites=Course.Prerequisites.from_json(json_data["prerequisites"]),
             optimized_prerequisites=optimized_prerequisites,
-            madgrades_data=madgrades_data,
-            enrollment_data={term: EnrollmentData.from_json(data) for term, data in json_data["enrollment_data"].items()}
+            cumulative_grade_data=cumulative_grade_data,
+            term_data={term: TermData.from_json(data) for term, data in json_data["term_data"].items()}
         )
 
     def to_dict(self):
@@ -147,8 +145,8 @@ class Course(JsonSerializable):
             "description": self.description,
             "prerequisites": self.prerequisites.to_dict(),
             "optimized_prerequisites": self.optimized_prerequisites.to_dict() if self.optimized_prerequisites else None,
-            "madgrades_data": self.madgrades_data.to_dict() if self.madgrades_data else None,
-            "enrollment_data": {term: data.to_dict() for term, data in self.enrollment_data.items()}
+            "cumulative_grade_data": self.cumulative_grade_data.to_dict() if self.cumulative_grade_data else None,
+            "term_data": {term: data.to_dict() for term, data in self.term_data.items()}
         }
 
     @classmethod
@@ -170,8 +168,15 @@ class Course(JsonSerializable):
         description = block.find("p", class_="courseblockdesc noindent").get_text(strip=True)
 
         cb_extras = block.find("div", class_="cb-extras")
-        basic_course = Course(course_reference, course_title, description, Course.Prerequisites("", set()), None, None,
-                              {})
+        basic_course = Course(
+            course_reference,
+            course_title,
+            description,
+            Course.Prerequisites("", set()),
+            None,
+            None,
+            {}
+        )
         if not cb_extras:
             return basic_course
 
@@ -190,7 +195,7 @@ class Course(JsonSerializable):
         requisites_courses.discard(course_reference)  # Remove self-reference
 
         course_prerequisite = Course.Prerequisites(requisites_text, requisites_courses)
-        return Course(course_reference, course_title, description, course_prerequisite, None, None, {})
+        return Course(course_reference, course_title, description, course_prerequisite, None, None,{})
 
     def determine_parent(self):
         parent = None
@@ -224,135 +229,3 @@ class Course(JsonSerializable):
     def __repr__(self):
         return self.get_identifier()
 
-
-class GradeData(JsonSerializable):
-
-    def __init__(self, total, a, ab, b, bc, c, d, f, satisfactory, unsatisfactory, credit, no_credit, passed,
-                 incomplete, no_work, not_reported, other):
-        self.total = total
-        self.a = a
-        self.ab = ab
-        self.b = b
-        self.bc = bc
-        self.c = c
-        self.d = d
-        self.f = f
-        self.satisfactory = satisfactory
-        self.unsatisfactory = unsatisfactory
-        self.credit = credit
-        self.no_credit = no_credit
-        self.passed = passed
-        self.incomplete = incomplete
-        self.no_work = no_work
-        self.not_reported = not_reported
-        self.other = other
-
-    @classmethod
-    def from_madgrades(cls, json_data) -> "GradeData":
-        return GradeData(
-            total=json_data["total"],
-            a=json_data["aCount"],
-            ab=json_data["abCount"],
-            b=json_data["bCount"],
-            bc=json_data["bcCount"],
-            c=json_data["cCount"],
-            d=json_data["dCount"],
-            f=json_data["fCount"],
-            satisfactory=json_data["sCount"],
-            unsatisfactory=json_data["uCount"],
-            credit=json_data["crCount"],
-            no_credit=json_data["nCount"],
-            passed=json_data["pCount"],
-            incomplete=json_data["iCount"],
-            no_work=json_data["nwCount"],
-            not_reported=json_data["nrCount"],
-            other=json_data["otherCount"]
-        )
-
-    @classmethod
-    def from_json(cls, json_data) -> "JsonSerializable":
-        return GradeData(
-            total=json_data["total"],
-            a=json_data["a"],
-            ab=json_data["ab"],
-            b=json_data["b"],
-            bc=json_data["bc"],
-            c=json_data["c"],
-            d=json_data["d"],
-            f=json_data["f"],
-            satisfactory=json_data["satisfactory"],
-            unsatisfactory=json_data["unsatisfactory"],
-            credit=json_data["credit"],
-            no_credit=json_data["no_credit"],
-            passed=json_data["passed"],
-            incomplete=json_data["incomplete"],
-            no_work=json_data["no_work"],
-            not_reported=json_data["not_reported"],
-            other=json_data["other"]
-        )
-
-    def to_dict(self):
-        return {
-            "total": self.total,
-            "a": self.a,
-            "ab": self.ab,
-            "b": self.b,
-            "bc": self.bc,
-            "c": self.c,
-            "d": self.d,
-            "f": self.f,
-            "satisfactory": self.satisfactory,
-            "unsatisfactory": self.unsatisfactory,
-            "credit": self.credit,
-            "no_credit": self.no_credit,
-            "passed": self.passed,
-            "incomplete": self.incomplete,
-            "no_work": self.no_work,
-            "not_reported": self.not_reported,
-            "other": self.other
-        }
-
-
-class MadgradesData(JsonSerializable):
-
-    def __init__(self, cumulative, by_term):
-        self.cumulative = cumulative
-        self.by_term = by_term
-
-    @classmethod
-    def from_json(cls, json_data) -> "MadgradesData":
-        return MadgradesData(
-            cumulative=GradeData.from_json(json_data["cumulative"]),
-            by_term={term: GradeData.from_json(data) for term, data in json_data["by_term"].items()}
-        )
-
-    def to_dict(self):
-        return {
-            "cumulative": self.cumulative.to_dict(),
-            "by_term": {term: data.to_dict() for term, data in self.by_term.items()}
-        }
-
-    @classmethod
-    async def from_madgrades_async(cls, session, url, madgrades_api_key, logger, attempts=3) -> "MadgradesData":
-        auth_header = {"Authorization": f"Token token={madgrades_api_key}"}
-
-        try:
-            async with session.get(url, headers=auth_header) as response:
-                data = await response.json()
-        except (JSONDecodeError, Exception) as e:
-            logger.warning(f"Failed to fetch madgrades data from {url}: {str(e)}")
-            if attempts > 0:
-                logger.info(f"Retrying {attempts} more times...")
-                await asyncio.sleep(1)
-                return await MadgradesData.from_madgrades_async(session, url, madgrades_api_key, logger, attempts - 1)
-
-        cumulative = GradeData.from_madgrades(data["cumulative"])
-        course_offerings = data["courseOfferings"]
-
-        by_term = {}
-        for offering in course_offerings:
-            term_code = offering["termCode"]
-            grade_data = GradeData.from_madgrades(offering["cumulative"])
-            by_term[term_code] = grade_data
-
-        return MadgradesData(cumulative=cumulative, by_term=by_term)
