+++ generation/enrollment_data.py
@@ -1,3 +1,6 @@
+import asyncio
+from json import JSONDecodeError
+
 from json_serializable import JsonSerializable
 from safe_parse import safe_int
 
@@ -39,8 +42,16 @@ class EnrollmentData(JsonSerializable):
         def __eq__(self, other):
             return self.name == other.name and self.abbreviation == other.abbreviation and self.url == other.url
 
-    def __init__(self, school, last_taught_term, typically_offered: str, credit_count: tuple[int, int],
-                 general_education: bool, ethnics_studies: bool, instructors):
+    def __init__(
+            self,
+            school: School | None,
+            last_taught_term: str,
+            typically_offered: str,
+            credit_count: tuple[int, int],
+            general_education: bool,
+            ethnics_studies: bool,
+            instructors,
+    ):
         self.school = school
         self.last_taught_term = last_taught_term
         self.typically_offered = typically_offered
@@ -87,3 +98,155 @@ class EnrollmentData(JsonSerializable):
             "ethnics_studies": self.ethnics_studies,
             "instructors": self.instructors
         }
+
+class GradeData(JsonSerializable):
+
+    def __init__(self, total, a, ab, b, bc, c, d, f, satisfactory, unsatisfactory, credit, no_credit, passed,
+                 incomplete, no_work, not_reported, other):
+        self.total = total
+        self.a = a
+        self.ab = ab
+        self.b = b
+        self.bc = bc
+        self.c = c
+        self.d = d
+        self.f = f
+        self.satisfactory = satisfactory
+        self.unsatisfactory = unsatisfactory
+        self.credit = credit
+        self.no_credit = no_credit
+        self.passed = passed
+        self.incomplete = incomplete
+        self.no_work = no_work
+        self.not_reported = not_reported
+        self.other = other
+
+    @classmethod
+    def from_madgrades(cls, json_data) -> "GradeData":
+        return GradeData(
+            total=json_data["total"],
+            a=json_data["aCount"],
+            ab=json_data["abCount"],
+            b=json_data["bCount"],
+            bc=json_data["bcCount"],
+            c=json_data["cCount"],
+            d=json_data["dCount"],
+            f=json_data["fCount"],
+            satisfactory=json_data["sCount"],
+            unsatisfactory=json_data["uCount"],
+            credit=json_data["crCount"],
+            no_credit=json_data["nCount"],
+            passed=json_data["pCount"],
+            incomplete=json_data["iCount"],
+            no_work=json_data["nwCount"],
+            not_reported=json_data["nrCount"],
+            other=json_data["otherCount"]
+        )
+
+    @classmethod
+    def from_json(cls, json_data) -> "GradeData":
+        return GradeData(
+            total=json_data["total"],
+            a=json_data["a"],
+            ab=json_data["ab"],
+            b=json_data["b"],
+            bc=json_data["bc"],
+            c=json_data["c"],
+            d=json_data["d"],
+            f=json_data["f"],
+            satisfactory=json_data["satisfactory"],
+            unsatisfactory=json_data["unsatisfactory"],
+            credit=json_data["credit"],
+            no_credit=json_data["no_credit"],
+            passed=json_data["passed"],
+            incomplete=json_data["incomplete"],
+            no_work=json_data["no_work"],
+            not_reported=json_data["not_reported"],
+            other=json_data["other"]
+        )
+
+    def to_dict(self):
+        return {
+            "total": self.total,
+            "a": self.a,
+            "ab": self.ab,
+            "b": self.b,
+            "bc": self.bc,
+            "c": self.c,
+            "d": self.d,
+            "f": self.f,
+            "satisfactory": self.satisfactory,
+            "unsatisfactory": self.unsatisfactory,
+            "credit": self.credit,
+            "no_credit": self.no_credit,
+            "passed": self.passed,
+            "incomplete": self.incomplete,
+            "no_work": self.no_work,
+            "not_reported": self.not_reported,
+            "other": self.other
+        }
+
+
+class MadgradesData:
+
+    def __init__(self, cumulative, by_term: dict[str, GradeData]):
+        self.cumulative = cumulative
+        self.by_term = by_term
+
+    @classmethod
+    async def from_madgrades_async(cls, session, url, madgrades_api_key, logger, attempts=3) -> "MadgradesData":
+        auth_header = {"Authorization": f"Token token={madgrades_api_key}"}
+
+        try:
+            async with session.get(url, headers=auth_header) as response:
+                data = await response.json()
+        except (JSONDecodeError, Exception) as e:
+            logger.warning(f"Failed to fetch madgrades data from {url}: {str(e)}")
+            if attempts > 0:
+                logger.info(f"Retrying {attempts} more times...")
+                await asyncio.sleep(1)
+                return await MadgradesData.from_madgrades_async(session, url, madgrades_api_key, logger, attempts - 1)
+
+        cumulative = GradeData.from_madgrades(data["cumulative"])
+        course_offerings = data["courseOfferings"]
+
+        by_term = {}
+        for offering in course_offerings:
+            term_code = str(offering["termCode"])
+            grade_data = GradeData.from_madgrades(offering["cumulative"])
+            by_term[term_code] = grade_data
+
+        return MadgradesData(cumulative=cumulative, by_term=by_term)
+
+
+class TermData(JsonSerializable):
+
+    def __init__(
+            self,
+            enrollment_data: EnrollmentData | None,
+            grade_data: GradeData | None,
+    ):
+        self.enrollment_data = enrollment_data
+        self.grade_data = grade_data
+
+
+    @classmethod
+    def from_json(cls, data) -> 'TermData':
+        enrollment_data = None
+        if data["enrollment_data"]:
+            enrollment_data = EnrollmentData.from_json(data["enrollment_data"])
+        grade_data = None
+        if data["grade_data"]:
+            grade_data = GradeData.from_json(data["grade_data"])
+        return TermData(
+            enrollment_data=enrollment_data,
+            grade_data=grade_data,
+        )
+
+    def to_dict(self) -> dict:
+        return {
+            "enrollment_data": self.enrollment_data.to_dict() if self.enrollment_data else None,
+            "grade_data": self.grade_data.to_dict() if self.grade_data else None,
+        }
+
+
